fix: IA:LiT asset URLs
Attempt #4: Apparently GitHub Pages involves Jekyll and Jekyll ignores
folders beginning with an underscore because it assumes they're "special
resources". Meanwhile NextJS thinks "_next" is an awesome name for a
folder containing all the static assets.

While the static assets are, of course, very special in their own way,
they are not special in the sense that Jeykll should ignore them, so
added ".nojeykll" file.