fix: adjusted pos of hole in "e" in SVG
Safari has clever ideas about text in SVGs, which caused issues with my
"good quality for evil intent" SVG.

I fixed that by converting the text to paths but, somehow, the hole in
the "e" in "evil" moved a pixel of so to the left.

This commit moves it back again.